requirements.txt
Flask==2.3.5
psycopg2-binary==2.9.7
redis==5.3.7